Task-42849: Transfer shareActivityOnSpaces rest service from Space rest resources to Activity rest resources
