+++ component/service/src/main/java/org/exoplatform/social/rest/impl/activity/ActivityRestResourcesV1.java
@@ -38,9 +38,11 @@ import org.exoplatform.social.core.activity.filter.ActivitySearchFilter;
 import org.exoplatform.social.core.activity.model.*;
 import org.exoplatform.social.core.identity.model.Identity;
 import org.exoplatform.social.core.identity.provider.OrganizationIdentityProvider;
+import org.exoplatform.social.core.identity.provider.SpaceIdentityProvider;
 import org.exoplatform.social.core.jpa.search.ActivitySearchConnector;
 import org.exoplatform.social.core.manager.ActivityManager;
 import org.exoplatform.social.core.manager.IdentityManager;
+import org.exoplatform.social.core.space.SpaceUtils;
 import org.exoplatform.social.core.space.model.Space;
 import org.exoplatform.social.core.space.spi.SpaceService;
 import org.exoplatform.social.core.storage.api.ActivityStorage;
@@ -50,6 +52,7 @@ import org.exoplatform.social.rest.api.RestUtils;
 import org.exoplatform.social.rest.entity.*;
 import org.exoplatform.social.service.rest.Util;
 import org.exoplatform.social.service.rest.api.VersionResources;
+import org.exoplatform.social.service.rest.api.models.SharedActivityRestIn;
 
 import io.swagger.annotations.*;
 
@@ -332,6 +335,64 @@ public class ActivityRestResourcesV1 implements ActivityRestResources {
     activityManager.saveComment(activity, comment);
     
     return EntityBuilder.getResponse(EntityBuilder.buildEntityFromComment(activityManager.getActivity(comment.getId()), uriInfo.getPath(), expand, false), uriInfo, RestUtils.getJsonMediaType(), Response.Status.OK);
+  }
+  
+  /**
+   * {@inheritDoc}
+   */
+  @POST
+  @Path("{id}/share")
+  @Produces(MediaType.APPLICATION_JSON)
+  @RolesAllowed("users")
+  @ApiOperation(value = "Shares a specific activity to specific spaces",
+                httpMethod = "POST",
+                response = Response.class,
+                notes = "This shares the given activity to the target spaces if the authenticated user has permissions to post to the target spaces")
+  @ApiResponses(value = {
+    @ApiResponse (code = 200, message = "Request fulfilled"),
+    @ApiResponse (code = 500, message = "Internal server error"),
+    @ApiResponse (code = 400, message = "Invalid query input") })
+  public Response shareActivityOnSpaces(@Context UriInfo uriInfo,
+                                      @ApiParam(value = "Activity id", required = true) @PathParam("id") String activityId,
+                                      @ApiParam(value = "Asking for a full representation of a specific subresource, ex: comments or likes", required = false) @QueryParam("expand") String expand,
+                                      @ApiParam(value = "Share target spaces", required = true) SharedActivityRestIn sharedActivityRestIn) throws Exception {
+    String authenticatedUser = ConversationState.getCurrent().getIdentity().getUserId();
+    ActivityManager activityManager = CommonsUtils.getService(ActivityManager.class);
+    if (activityManager.getActivity(activityId) == null) {
+      return Response.status(Response.Status.NOT_FOUND).build();
+    }
+
+    if (sharedActivityRestIn == null || sharedActivityRestIn.getTargetSpaces() == null || sharedActivityRestIn.getTargetSpaces().isEmpty() || sharedActivityRestIn.getType() == null) {
+      return Response.status(Response.Status.BAD_REQUEST).build();
+    }
+    
+    Identity authenticatedUserIdentity = CommonsUtils.getService(IdentityManager.class).getOrCreateIdentity(OrganizationIdentityProvider.NAME, authenticatedUser);
+    List<ActivityEntity> sharedActivitiesEntities = new ArrayList<ActivityEntity>();
+    for (String targetSpaceName : sharedActivityRestIn.getTargetSpaces()) {
+      Space targetSpace = CommonsUtils.getService(SpaceService.class).getSpaceByPrettyName(targetSpaceName);
+      if (SpaceUtils.isSpaceManagerOrSuperManager(authenticatedUser, targetSpace.getGroupId()) || (CommonsUtils.getService(SpaceService.class).isMember(targetSpace, authenticatedUser) && SpaceUtils.isRedactor(authenticatedUser, targetSpace.getGroupId()))) {
+        // create activity
+        ExoSocialActivity sharedActivity = new ExoSocialActivityImpl();
+        sharedActivity.setTitle(sharedActivityRestIn.getTitle());
+        sharedActivity.setType(sharedActivityRestIn.getType());
+        sharedActivity.setUserId(authenticatedUserIdentity.getId());
+        Map<String, String> templateParams = new HashMap<>();
+        templateParams.put("originalActivityId", activityId); 
+        sharedActivity.setTemplateParams(templateParams);
+        Identity targetSpaceIdentity = CommonsUtils.getService(IdentityManager.class).getOrCreateIdentity(SpaceIdentityProvider.NAME, targetSpaceName);
+        if (targetSpaceIdentity != null) {
+          CommonsUtils.getService(ActivityManager.class).saveActivityNoReturn(targetSpaceIdentity, sharedActivity);
+          ActivityEntity sharedActivityEntity = EntityBuilder.buildEntityFromActivity(sharedActivity, uriInfo.getPath(), expand);
+          sharedActivitiesEntities.add(sharedActivityEntity);
+          LOG.info("service=activity operation=share parameters=\"activity_type:{},activity_id:{},space_id:{},user_id:{}\"",
+                   sharedActivity.getType(),
+                   sharedActivity.getId(),
+                   targetSpace.getId(),
+                   sharedActivity.getPosterId());
+        }
+      }
+    }
+    return EntityBuilder.getResponse(sharedActivitiesEntities, uriInfo, RestUtils.getJsonMediaType(), Response.Status.OK);
   }
   
   @GET
